+++ public/common/TracyQueue.hpp
@@ -414,7 +414,7 @@ struct QueueGpuNewContext
     int64_t gpuTime;
     uint32_t thread;
     float period;
-    uint8_t context;
+    uint16_t context;
     GpuContextFlags flags;
     GpuContextType type;
 };
@@ -424,7 +424,7 @@ struct QueueGpuZoneBeginLean
     int64_t cpuTime;
     uint32_t thread;
     uint16_t queryId;
-    uint8_t context;
+    uint16_t context;
 };
 
 struct QueueGpuZoneBegin : public QueueGpuZoneBeginLean
@@ -437,14 +437,14 @@ struct QueueGpuZoneEnd
     int64_t cpuTime;
     uint32_t thread;
     uint16_t queryId;
-    uint8_t context;
+    uint16_t context;
 };
 
 struct QueueGpuTime
 {
     int64_t gpuTime;
     uint16_t queryId;
-    uint8_t context;
+    uint16_t context;
 };
 
 struct QueueGpuCalibration
@@ -452,19 +452,19 @@ struct QueueGpuCalibration
     int64_t gpuTime;
     int64_t cpuTime;
     int64_t cpuDelta;
-    uint8_t context;
+    uint16_t context;
 };
 
 struct QueueGpuTimeSync
 {
     int64_t gpuTime;
     int64_t cpuTime;
-    uint8_t context;
+    uint16_t context;
 };
     
 struct QueueGpuContextName
 {
-    uint8_t context;
+    uint16_t context;
 };
 
 struct QueueGpuContextNameFat : public QueueGpuContextName
